+++ src/components/Votes/VoteDetails/InfoText/InfoText.css
