+++ src/components/Votes/VoteDetails/InfoText/Infotext.tsx
