+++ src/components/Votes/VoteDetails/Middlebar/Middlebar.css
