+++ src/components/Votes/VoteDetails/Preview/Preview.css
