+++ src/components/Votes/VoteDetails/Preview/Preview.tsx
