+++ src/components/Votes/VoteDetails/Table/Content/Columns/Column/Column.css
