+++ src/components/Votes/VoteDetails/Table/Content/Columns/Columns.css
