+++ src/components/Votes/VoteDetails/Table/Content/Columns/TotalColumn/TotalColumn.tsx
