+++ src/components/Votes/VoteDetails/Table/Content/Content.css
