+++ src/components/Votes/VoteDetails/Table/NumberChips/NumberChip/NumberChip.css
