+++ src/components/Votes/VoteDetails/Table/NumberChips/NumberChips.css
