+++ src/components/Votes/VoteDetails/Table/PartyChips/PartyChip/PartyChip.tsx
