+++ src/components/Votes/VoteDetails/Table/PositioningColumns/PositioningColumns.css
