+++ src/components/Votes/VoteDetails/Table/Table.css
