+++ src/components/Votes/VoteDetails/Table/Table.tsx
