+++ src/components/Votes/VoteDetails/VoteDetails.tsx
