+++ src/components/Votes/VoteDetails/VotesDetailsToolbar/VoteDetailsToolbar.css
